Refactor create test data methods to helper_test.py
